refactor: remove version field from all messages
- Strip redundant version field from Request, RenderResult, Update, Hello, HelloAck, StatusRequest, StatusResponse
- Add MAX_MESSAGE_SIZE and refactor codec to use fill_buf
- Add MessageTooLarge error variant; const-assert PROTOCOL_VERSION_BYTES
- Validate SourceDef at deserialization time (exactly-one source, safe paths)
- Add is_safe_relative_module_path helper
- Introduce JoinSet-based task lifecycle for handlers and workers
- Add ActiveConnectionGuard drop guard for connection stats
- Add SlowWorkClaim enum unifying cache-hit and inflight logic
- Add async GitStatusFuture with tokio::process git status
- Expand ModuleDependencyInputs with source_files and uses_command_path
- Add signal_process, retry negotiation, and escaped tab protocol in init.zsh
- Extract inline test modules to separate files under each crate
