+++ crates/core/src/module/git.rs
@@ -1,7 +1,9 @@
 //! Git module — displays git branch and working tree status.
 
 use std::{
+    future::Future,
     path::{Path, PathBuf},
+    pin::Pin,
     process::Command,
 };
 
@@ -93,6 +95,10 @@ pub struct GitStatus {
     pub state: Option<GitOperationState>,
 }
 
+/// Boxed future returned by async git status providers.
+pub type GitStatusFuture =
+    Pin<Box<dyn Future<Output = Result<Option<GitStatus>, GitError>> + Send + 'static>>;
+
 /// Provides git repository information.
 pub trait GitProvider: sealed::Sealed {
     /// Query the git status of the repository at `cwd`.
@@ -107,6 +113,24 @@ pub trait GitProvider: sealed::Sealed {
     ///
     /// Returns [`GitError`] if the git command cannot be executed.
     fn status(&self, cwd: &Path, path_env: Option<&str>) -> Result<Option<GitStatus>, GitError>;
+
+    /// Async daemon-facing status query.
+    ///
+    /// The default implementation delegates synchronous providers to a
+    /// blocking thread. [`CommandGitProvider`] overrides this to run the git
+    /// process with async I/O and `kill_on_drop`.
+    fn status_async(self, cwd: PathBuf, path_env: Option<String>) -> GitStatusFuture
+    where
+        Self: Sized + Send + 'static,
+    {
+        Box::pin(async move {
+            tokio::task::spawn_blocking(move || self.status(&cwd, path_env.as_deref()))
+                .await
+                .map_err(|source| {
+                    GitError::Command(std::io::Error::other(format!("join error: {source}")))
+                })?
+        })
+    }
 }
 
 /// [`GitProvider`] that shells out to the `git` command.
@@ -141,6 +165,41 @@ impl GitProvider for CommandGitProvider {
 
         Ok(Some(status))
     }
+
+    fn status_async(self, cwd: PathBuf, path_env: Option<String>) -> GitStatusFuture {
+        Box::pin(async move { command_git_status_async(&cwd, path_env.as_deref()).await })
+    }
+}
+
+async fn command_git_status_async(
+    cwd: &Path,
+    path_env: Option<&str>,
+) -> Result<Option<GitStatus>, GitError> {
+    let mut cmd = tokio::process::Command::new("git");
+    cmd.kill_on_drop(true)
+        .args(["status", "--porcelain=v2", "--branch", "--show-stash"])
+        .current_dir(cwd)
+        .stderr(std::process::Stdio::null());
+    if let Some(path) = path_env {
+        cmd.env("PATH", path);
+    }
+    let output = cmd.output().await.map_err(GitError::Command)?;
+
+    if !output.status.success() {
+        return Ok(None);
+    }
+
+    let stdout = String::from_utf8_lossy(&output.stdout);
+    let mut status = parse_porcelain_v2(&stdout);
+
+    let state_cwd = cwd.to_path_buf();
+    status.state = tokio::task::spawn_blocking(move || {
+        find_git_dir(&state_cwd).and_then(|git_dir| detect_git_state(&git_dir))
+    })
+    .await
+    .map_err(|source| GitError::Command(std::io::Error::other(format!("join error: {source}"))))?;
+
+    Ok(Some(status))
 }
 
 /// Bundled style configuration for git output rendering.
@@ -196,6 +255,10 @@ impl<G> GitModule<G> {
 }
 
 impl<G: GitProvider> GitModule<G> {
+    pub(crate) fn render_status(&self, status: &GitStatus) -> Option<ModuleOutput> {
+        render_status_with_styles(status, &self.styles)
+    }
+
     /// Renders git status for the given working directory.
     ///
     /// This is the core implementation used by both [`Module::render`] and
@@ -209,12 +272,19 @@ impl<G: GitProvider> GitModule<G> {
                 return None;
             }
         };
-        let content = format_git_output(&status, &self.styles);
-        if content.is_empty() {
-            return None;
-        }
-        Some(ModuleOutput { content })
+        self.render_status(&status)
     }
+}
+
+pub(crate) fn render_status_with_styles(
+    status: &GitStatus,
+    styles: &GitStyles,
+) -> Option<ModuleOutput> {
+    let content = format_git_output(status, styles);
+    if content.is_empty() {
+        return None;
+    }
+    Some(ModuleOutput { content })
 }
 
 impl<G: GitProvider> sealed::Sealed for GitModule<G> {}
@@ -392,8 +462,8 @@ fn parse_changed_entry(line: &str, status: &mut GitStatus) {
 const DETACHED_OID_PREFIX_LEN: usize = 7;
 
 fn short_commit_prefix(full_oid: &str) -> &str {
-    let end = full_oid.len().min(DETACHED_OID_PREFIX_LEN);
-    full_oid.get(..end).unwrap_or("")
+    // Git OIDs are ASCII hex, so byte-slicing is safe up to len.
+    &full_oid[..full_oid.len().min(DETACHED_OID_PREFIX_LEN)]
 }
 
 fn write_state(buf: &mut String, op_state: &GitOperationState) {
@@ -417,10 +487,7 @@ fn format_git_output(status: &GitStatus, styles: &GitStyles) -> String {
         let prefix = short_commit_prefix(oid);
         if !prefix.is_empty() {
             out.push_str(&styles.branch.paint_with("HEAD ", styles.color_map));
-            let mut paren = String::with_capacity(prefix.len() + 2);
-            paren.push('(');
-            paren.push_str(prefix);
-            paren.push(')');
+            let paren = format!("({prefix})");
             out.push_str(&styles.detached_hash.paint_with(&paren, styles.color_map));
         }
     }
@@ -480,873 +547,4 @@ fn format_git_output(status: &GitStatus, styles: &GitStyles) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::path::Path;
-
-    use super::*;
-    use crate::{render::layout::display_width, test_utils::contains_style_sequence};
-
-    // -- GitState Display tests --
-
-    #[test]
-    fn test_git_state_display_rebase() {
-        assert_eq!(GitState::Rebase.to_string(), "REBASING");
-    }
-
-    #[test]
-    fn test_git_state_display_am() {
-        assert_eq!(GitState::Am.to_string(), "AM");
-    }
-
-    #[test]
-    fn test_git_state_display_merge() {
-        assert_eq!(GitState::Merge.to_string(), "MERGING");
-    }
-
-    #[test]
-    fn test_git_state_display_cherry_pick() {
-        assert_eq!(GitState::CherryPick.to_string(), "CHERRY-PICKING");
-    }
-
-    #[test]
-    fn test_git_state_display_revert() {
-        assert_eq!(GitState::Revert.to_string(), "REVERTING");
-    }
-
-    #[test]
-    fn test_git_state_display_bisect() {
-        assert_eq!(GitState::Bisect.to_string(), "BISECTING");
-    }
-
-    // -- find_git_dir tests --
-
-    #[test]
-    fn test_find_git_dir_normal_repo() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::create_dir(dir.path().join(".git"))?;
-        let result = find_git_dir(dir.path());
-        assert_eq!(result, Some(dir.path().join(".git")));
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_git_dir_subdirectory() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::create_dir(dir.path().join(".git"))?;
-        let sub = dir.path().join("src").join("deep");
-        std::fs::create_dir_all(&sub)?;
-        let result = find_git_dir(&sub);
-        assert_eq!(result, Some(dir.path().join(".git")));
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_git_dir_worktree() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let gitdir_target = dir.path().join("actual-gitdir");
-        std::fs::create_dir(&gitdir_target)?;
-        let worktree = dir.path().join("worktree");
-        std::fs::create_dir(&worktree)?;
-        std::fs::write(
-            worktree.join(".git"),
-            format!("gitdir: {}", gitdir_target.display()),
-        )?;
-        let result = find_git_dir(&worktree);
-        assert_eq!(result, Some(gitdir_target));
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_git_dir_worktree_relative() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let gitdir_target = dir.path().join("actual-gitdir");
-        std::fs::create_dir(&gitdir_target)?;
-        let worktree = dir.path().join("worktree");
-        std::fs::create_dir(&worktree)?;
-        std::fs::write(worktree.join(".git"), "gitdir: ../actual-gitdir\n")?;
-        let result = find_git_dir(&worktree);
-        assert!(result.is_some(), "should resolve relative gitdir pointer");
-        assert!(
-            result
-                .as_ref()
-                .is_some_and(|p| p.ends_with("actual-gitdir")),
-            "resolved path should end with actual-gitdir: {result:?}",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_find_git_dir_not_a_repo() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let result = find_git_dir(dir.path());
-        // tempdir is under /tmp or similar — may find system .git if any; safest
-        // is to verify that the returned path (if any) is not inside our tempdir.
-        if let Some(ref p) = result {
-            assert!(
-                !p.starts_with(dir.path()),
-                "should not find .git inside our tempdir: {p:?}",
-            );
-        }
-        Ok(())
-    }
-
-    // -- detect_git_state tests --
-
-    #[test]
-    fn test_detect_rebase_merge_with_progress() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let rebase = dir.path().join("rebase-merge");
-        std::fs::create_dir(&rebase)?;
-        std::fs::write(rebase.join("msgnum"), "3\n")?;
-        std::fs::write(rebase.join("end"), "7\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState {
-                state: GitState::Rebase,
-                step: Some(3),
-                total: Some(7),
-            }),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_rebase_merge_without_progress() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::create_dir(dir.path().join("rebase-merge"))?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState {
-                state: GitState::Rebase,
-                step: None,
-                total: None,
-            }),
-            "rebase-merge dir without msgnum/end should have None step/total",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_rebase_apply() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let rebase = dir.path().join("rebase-apply");
-        std::fs::create_dir(&rebase)?;
-        std::fs::write(rebase.join("next"), "2\n")?;
-        std::fs::write(rebase.join("last"), "5\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState {
-                state: GitState::Rebase,
-                step: Some(2),
-                total: Some(5),
-            }),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_am() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let rebase = dir.path().join("rebase-apply");
-        std::fs::create_dir(&rebase)?;
-        std::fs::write(rebase.join("applying"), "")?;
-        std::fs::write(rebase.join("next"), "1\n")?;
-        std::fs::write(rebase.join("last"), "3\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState {
-                state: GitState::Am,
-                step: Some(1),
-                total: Some(3),
-            }),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_merge() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::write(dir.path().join("MERGE_HEAD"), "abc123\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState::without_progress(GitState::Merge)),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_cherry_pick() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::write(dir.path().join("CHERRY_PICK_HEAD"), "abc123\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState::without_progress(GitState::CherryPick)),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_revert() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::write(dir.path().join("REVERT_HEAD"), "abc123\n")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState::without_progress(GitState::Revert)),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_bisect() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::write(dir.path().join("BISECT_LOG"), "")?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(
-            result,
-            Some(GitOperationState::without_progress(GitState::Bisect)),
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_no_state() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let result = detect_git_state(dir.path());
-        assert_eq!(result, None);
-        Ok(())
-    }
-
-    #[test]
-    fn test_detect_priority_rebase_over_merge() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        std::fs::create_dir(dir.path().join("rebase-merge"))?;
-        std::fs::write(dir.path().join("MERGE_HEAD"), "abc123\n")?;
-        let result = detect_git_state(dir.path());
-        assert!(
-            result.is_some_and(|s| s.state == GitState::Rebase),
-            "rebase should take priority over merge: {result:?}",
-        );
-        Ok(())
-    }
-
-    fn default_styles() -> GitStyles {
-        GitStyles::default()
-    }
-
-    // -- Parsing tests --
-
-    #[test]
-    fn test_parse_porcelain_v2_branch_and_counts() {
-        let output = "\
-# branch.oid abc123def456
-# branch.head main
-# branch.ab +1 -2
-1 M. N... 000000 000000 abc123 def456 modified.rs
-1 .M N... 000000 000000 abc123 def456 worktree.rs
-? untracked.txt
-";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.branch, Some("main".to_owned()));
-        assert_eq!(status.ahead, 1);
-        assert_eq!(status.behind, 2);
-        assert_eq!(status.staged, 1);
-        assert_eq!(status.modified, 1);
-        assert_eq!(status.untracked, 1);
-        assert_eq!(status.conflicted, 0);
-        assert_eq!(
-            status.head_oid,
-            Some("abc123def456".to_owned()),
-            "full oid from porcelain"
-        );
-    }
-
-    #[test]
-    fn test_parse_porcelain_v2_detached_head() {
-        let output = "# branch.oid abc123\n# branch.head (detached)\n";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.branch, None);
-        assert_eq!(status.head_oid, Some("abc123".to_owned()));
-    }
-
-    #[test]
-    fn test_parse_porcelain_v2_staged_and_modified() {
-        let output = "# branch.head feature\n1 MM N... 000000 000000 abc123 def456 both.rs\n";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.staged, 1);
-        assert_eq!(status.modified, 1);
-    }
-
-    #[test]
-    fn test_parse_porcelain_v2_conflicted() {
-        let output =
-            "# branch.head main\nu UU N... 000000 000000 000000 abc123 def456 ghi789 conflict.rs\n";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.conflicted, 1);
-    }
-
-    #[test]
-    fn test_parse_porcelain_v2_rename_entry() {
-        let output =
-            "# branch.head main\n2 R. N... 000000 000000 abc123 def456 R100 new.rs\told.rs\n";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.staged, 1);
-        assert_eq!(status.modified, 0);
-    }
-
-    #[test]
-    fn test_parse_porcelain_v2_empty_output() {
-        let status = parse_porcelain_v2("");
-        assert_eq!(status, GitStatus::default());
-    }
-
-    // -- Format tests --
-
-    #[test]
-    fn test_format_git_output_branch_only() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert_eq!(display_width(&output), 4, "visible width: {output:?}");
-        assert!(output.contains("main"), "should contain branch name");
-        assert!(
-            contains_style_sequence(&output, &[1, 35]),
-            "branch should be bold magenta"
-        );
-        // No indicators → display width is just the branch name
-        assert_eq!(
-            display_width(&output),
-            display_width("main"),
-            "no extra content when no status"
-        );
-    }
-
-    #[test]
-    fn test_format_git_output_bracket_indicators() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            staged: 2,
-            modified: 1,
-            untracked: 3,
-            ahead: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        // "main [!+?⇡]" = 4 + 1 + 6 = 11 visible chars
-        assert_eq!(display_width(&output), 11, "visible width: {output:?}");
-        assert!(output.contains("main"), "should contain branch");
-        assert!(
-            output.contains("[!+?⇡]"),
-            "should contain bracketed indicators: {output:?}"
-        );
-        assert!(
-            contains_style_sequence(&output, &[1, 31]),
-            "brackets should be bold red: {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_git_output_detached_clean_shows_short_oid() {
-        let status = GitStatus {
-            branch: None,
-            head_oid: Some("abcdef0123456789abcdef0123456789abcd".to_owned()),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert_eq!(display_width(&output), 14, "visible width: {output:?}");
-        assert!(
-            output.contains("HEAD ") && output.contains("(abcdef0)"),
-            "detached label should be HEAD (short oid); zsh escapes may split segments: {output:?}"
-        );
-        assert!(
-            contains_style_sequence(&output, &[1, 35]),
-            "HEAD should use branch style bold magenta: {output:?}"
-        );
-        assert!(
-            contains_style_sequence(&output, &[2, 32])
-                || contains_style_sequence(&output, &[32, 2]),
-            "(hash) should use dimmed green: {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_git_output_detached_with_indicators() {
-        let status = GitStatus {
-            branch: None,
-            head_oid: Some("deadbeef".to_owned()),
-            modified: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        let clean = strip_ansi_and_zsh(&output);
-        assert_eq!(
-            clean, "HEAD (deadbee) [!]",
-            "short oid shorter than 7 uses full hash inside parens: {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_git_output_no_branch() {
-        let status = GitStatus {
-            branch: None,
-            staged: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        // "[+]" = 3 visible chars
-        assert_eq!(display_width(&output), 3, "visible width: {output:?}");
-        assert!(
-            output.contains("[+]"),
-            "should contain bracketed staged indicator: {output:?}"
-        );
-        assert!(
-            contains_style_sequence(&output, &[1, 31]),
-            "brackets should be bold red: {output:?}"
-        );
-    }
-
-    // -- Mock provider tests --
-
-    struct MockGitProvider {
-        result: Option<GitStatus>,
-    }
-
-    impl sealed::Sealed for MockGitProvider {}
-
-    impl GitProvider for MockGitProvider {
-        fn status(
-            &self,
-            _cwd: &Path,
-            _path_env: Option<&str>,
-        ) -> Result<Option<GitStatus>, GitError> {
-            Ok(self.result.clone())
-        }
-    }
-
-    fn make_ctx() -> RenderContext<'static> {
-        RenderContext {
-            cwd: Path::new("/tmp"),
-            home_dir: Path::new("/Users/testuser"),
-            last_exit_code: 0,
-            duration_ms: None,
-            keymap: "main",
-            cols: 80,
-        }
-    }
-
-    #[test]
-    fn test_module_not_a_repo_returns_none() {
-        let module = GitModule::new(MockGitProvider { result: None });
-        let ctx = make_ctx();
-        assert!(module.render(&ctx).is_none());
-    }
-
-    #[test]
-    fn test_module_staged_changes() {
-        let module = GitModule::new(MockGitProvider {
-            result: Some(GitStatus {
-                branch: Some("main".to_owned()),
-                staged: 2,
-                ..GitStatus::default()
-            }),
-        });
-        let ctx = make_ctx();
-        let output = module.render(&ctx);
-        assert!(output.is_some());
-        let content = output.map(|o| o.content).unwrap_or_default();
-        assert!(
-            content.contains("[+]"),
-            "expected bracketed staged indicator in: {content}"
-        );
-    }
-
-    #[test]
-    fn test_module_speed_is_slow() {
-        let module = GitModule::new(MockGitProvider { result: None });
-        assert_eq!(module.speed(), ModuleSpeed::Slow);
-    }
-
-    // -- Integration test with real git --
-
-    #[test]
-    fn test_module_real_git_repo_with_staged_file() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let path = dir.path();
-
-        // Initialize a git repo
-        let init = Command::new("git")
-            .args(["init", "-b", "main"])
-            .current_dir(path)
-            .output()?;
-        assert!(init.status.success(), "git init failed");
-
-        // Configure git identity (needed in CI)
-        Command::new("git")
-            .args(["config", "user.name", "test"])
-            .current_dir(path)
-            .output()?;
-        Command::new("git")
-            .args(["config", "user.email", "test@test.com"])
-            .current_dir(path)
-            .output()?;
-
-        // Create and stage a file
-        std::fs::write(path.join("hello.txt"), "hello")?;
-        let add = Command::new("git")
-            .args(["add", "hello.txt"])
-            .current_dir(path)
-            .output()?;
-        assert!(add.status.success(), "git add failed");
-
-        // Query via CommandGitProvider
-        let provider = CommandGitProvider;
-        let status = provider.status(path, None)?;
-        let status = status.as_ref();
-        assert!(status.is_some(), "should detect git repo");
-        assert!(
-            status.is_some_and(|s| s.staged > 0),
-            "should have staged files"
-        );
-
-        // Query via GitModule
-        let module = GitModule::new(CommandGitProvider);
-        let ctx = RenderContext {
-            cwd: path,
-            home_dir: Path::new("/nonexistent"),
-            last_exit_code: 0,
-            duration_ms: None,
-            keymap: "main",
-            cols: 80,
-        };
-        let output = module.render(&ctx);
-        assert!(output.is_some(), "git module should produce output");
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_module_not_a_git_repo() -> Result<(), Box<dyn std::error::Error>> {
-        let dir = tempfile::tempdir()?;
-        let provider = CommandGitProvider;
-        let status = provider.status(dir.path(), None)?;
-        assert!(status.is_none(), "non-git dir should return None");
-        Ok(())
-    }
-
-    // -- Starship-compatible indicator tests --
-
-    #[test]
-    fn test_format_conflict_uses_equals_sign() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            conflicted: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(
-            output.contains("[=]"),
-            "conflict should use '=' not '~': {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_stash_indicator() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            stashed: 3,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(output.contains("[$]"), "stash should show '$': {output:?}");
-    }
-
-    #[test]
-    fn test_format_deleted_indicator() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            deleted: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(
-            output.contains("[✘]"),
-            "deleted should show '✘': {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_renamed_indicator() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            renamed: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(
-            output.contains("[»]"),
-            "renamed should show '»': {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_diverged_indicator() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            ahead: 2,
-            behind: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(
-            output.contains('⇕'),
-            "diverged (ahead+behind) should show '⇕': {output:?}"
-        );
-        assert!(
-            !output.contains('⇡'),
-            "diverged should not show separate '⇡': {output:?}"
-        );
-        assert!(
-            !output.contains('⇣'),
-            "diverged should not show separate '⇣': {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_indicator_order() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            conflicted: 1,
-            stashed: 1,
-            deleted: 1,
-            renamed: 1,
-            modified: 1,
-            staged: 1,
-            untracked: 1,
-            ahead: 1,
-            behind: 0,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        // Strip all ANSI/zsh escapes to get visible text
-        let clean = strip_ansi_and_zsh(&output);
-        // Expected visible: "main [=$✘»!+?⇡]"
-        assert_eq!(
-            clean, "main [=$✘»!+?⇡]",
-            "indicators should be in Starship order: {output:?}"
-        );
-    }
-
-    #[test]
-    fn test_format_git_output_custom_styles() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            modified: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(
-            &status,
-            &GitStyles {
-                branch: Style::new().fg(Color::Cyan),
-                detached_hash: Style::new().fg(Color::Green),
-                indicator: Style::new().fg(Color::Yellow),
-                color_map: ColorMap {
-                    cyan: 96,
-                    green: 32,
-                    yellow: 93,
-                    ..ColorMap::default()
-                },
-                ..default_styles()
-            },
-        );
-        assert!(
-            output.contains("\x1b[96m"),
-            "branch should use remapped cyan: {output:?}"
-        );
-        assert!(
-            output.contains("\x1b[93m"),
-            "indicators should use remapped yellow: {output:?}"
-        );
-    }
-
-    /// Strip ANSI escape sequences and zsh `%{{..%}}` wrappers.
-    fn strip_ansi_and_zsh(s: &str) -> String {
-        let mut result = String::with_capacity(s.len());
-        let mut chars = s.chars().peekable();
-        while let Some(c) = chars.next() {
-            if c == '%' && chars.peek() == Some(&'{') {
-                // Skip %{...%}
-                chars.next(); // consume '{'
-                while let Some(inner) = chars.next() {
-                    if inner == '%' && chars.peek() == Some(&'}') {
-                        chars.next(); // consume '}'
-                        break;
-                    }
-                }
-            } else {
-                result.push(c);
-            }
-        }
-        result
-    }
-
-    // -- write_state tests --
-
-    fn state_to_string(op_state: &GitOperationState) -> String {
-        let mut buf = String::new();
-        write_state(&mut buf, op_state);
-        buf
-    }
-
-    #[test]
-    fn test_write_state_progress() {
-        let state = GitOperationState {
-            state: GitState::Rebase,
-            step: Some(2),
-            total: Some(5),
-        };
-        assert_eq!(state_to_string(&state), "(REBASING 2/5)");
-    }
-
-    #[test]
-    fn test_write_state_no_progress() {
-        let state = GitOperationState::without_progress(GitState::Merge);
-        assert_eq!(state_to_string(&state), "(MERGING)");
-    }
-
-    #[test]
-    fn test_write_state_partial_progress_falls_back() {
-        let state = GitOperationState {
-            state: GitState::Rebase,
-            step: Some(3),
-            total: None,
-        };
-        assert_eq!(
-            state_to_string(&state),
-            "(REBASING)",
-            "partial progress should fall back to no-progress display",
-        );
-    }
-
-    // -- format_git_output with state tests --
-
-    #[test]
-    fn test_format_git_output_rebase_state() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            state: Some(GitOperationState {
-                state: GitState::Rebase,
-                step: Some(2),
-                total: Some(5),
-            }),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        let clean = strip_ansi_and_zsh(&output);
-        assert_eq!(clean, "main (REBASING 2/5)");
-    }
-
-    #[test]
-    fn test_format_git_output_merge_state() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            state: Some(GitOperationState::without_progress(GitState::Merge)),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        let clean = strip_ansi_and_zsh(&output);
-        assert_eq!(clean, "main (MERGING)");
-    }
-
-    #[test]
-    fn test_format_git_output_state_indicators() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            state: Some(GitOperationState {
-                state: GitState::Rebase,
-                step: Some(2),
-                total: Some(5),
-            }),
-            modified: 1,
-            staged: 1,
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        let clean = strip_ansi_and_zsh(&output);
-        assert_eq!(clean, "main (REBASING 2/5) [!+]");
-    }
-
-    #[test]
-    fn test_format_git_output_detached_state() {
-        let status = GitStatus {
-            branch: None,
-            head_oid: Some("abcdef0123456789".to_owned()),
-            state: Some(GitOperationState::without_progress(GitState::CherryPick)),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        let clean = strip_ansi_and_zsh(&output);
-        assert_eq!(clean, "HEAD (abcdef0) (CHERRY-PICKING)");
-    }
-
-    #[test]
-    fn test_format_git_output_state_style() {
-        let status = GitStatus {
-            branch: Some("main".to_owned()),
-            state: Some(GitOperationState::without_progress(GitState::Merge)),
-            ..GitStatus::default()
-        };
-        let output = format_git_output(&status, &default_styles());
-        assert!(
-            contains_style_sequence(&output, &[1, 33]),
-            "state should be bold yellow: {output:?}",
-        );
-    }
-
-    #[test]
-    fn test_parse_stash_count() {
-        let output = "\
-# branch.head main
-# stash 5
-";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.stashed, 5);
-    }
-
-    #[test]
-    fn test_parse_deleted_file() {
-        let output = "\
-# branch.head main
-1 D. N... 100644 000000 000000 abc123 000000 deleted.rs
-";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.deleted, 1, "index delete should be tracked");
-    }
-
-    #[test]
-    fn test_parse_worktree_deleted_file() {
-        let output = "\
-# branch.head main
-1 .D N... 100644 100644 000000 abc123 def456 deleted.rs
-";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.deleted, 1, "worktree delete should be tracked");
-    }
-
-    #[test]
-    fn test_parse_renamed_file() {
-        let output = "\
-# branch.head main
-2 R. N... 100644 100644 100644 abc123 def456 R100 new.rs\told.rs
-";
-        let status = parse_porcelain_v2(output);
-        assert_eq!(status.renamed, 1, "rename should be tracked");
-    }
-}
+mod tests;
